--- miniprogram/alicloud/.gitignore
@@ -1,2 +1,0 @@
-/releases.js
-/dev.js
