+++ miniprogram/app.json
@@ -43,8 +43,8 @@
         "text": ""
       },
       {
-        "text": "",
-        "pagePath": "pages/giftOut/index"
+        "pagePath": "pages/giftOut/index",
+        "text": ""
       },
       {
         "pagePath": "pages/friend/index",
