+++ miniprogram/package.json
@@ -9,7 +9,6 @@
   "author": "chadwuo",
   "license": "ISC",
   "dependencies": {
-    "@alicloud/mpserverless-sdk": "^3.1.3",
     "@vant/weapp": "^1.10.15",
     "dayjs": "^1.10.7",
     "flyio": "^0.6.14",
