+++ miniprogram/pages/friend/edit/index.js
@@ -1,5 +1,5 @@
 // pages/friend/edit/index.js
-const friendService = require("@/alicloud/services/friend");
+const friendService = require("@/services/friend");
 const app = getApp();
 Page({
   /**
