+++ miniprogram/pages/giftOut/index.js
@@ -1,5 +1,5 @@
 // pages/giftOut/index.js
-const giftOutService = require('@/alicloud/services/giftOut');
+const giftOutService = require('@/services/giftOut');
 const app = getApp();
 Page({
   /**
