+++ miniprogram/pages/giftReceive/edit/index.js
@@ -1,5 +1,5 @@
 // pages/giftReceive/edit/index.js
-const giftReceiveService = require("@/alicloud/services/giftReceive");
+const giftReceiveService = require("@/services/giftReceive");
 const app = getApp();
 Page({
   /**
