+++ miniprogram/pages/start/index.js
@@ -1,6 +1,6 @@
 // pages/start/index.js
 const app = getApp();
-import { getUserInfo } from "@/alicloud/services/user";
+import { getUserInfo } from "@/services/user";
 
 Page({
   /**
