+++ miniprogram/services/friend.js
