+++ miniprogram/services/sponsor.js
